+++ bindings/nostr-sdk-ffi/src/client/mod.rs
@@ -524,8 +524,6 @@ impl Client {
     }
 
     /// Send a Zap!
-    ///
-    /// This method automatically create a split zap to support Rust Nostr development.
     pub fn zap(
         &self,
         to: Arc<ZapEntity>,